Initial commit.
The plan for this is that I am going to write the functionality as a CLI program. Once that is complete, I am going to add a GUI element and refactor the code to use the GUI element. I suspect that I will have to do less exception handling with the GUI version, but I want a proof of concept of the CLI one.